+++ src/main/java/com/pintogether/backend/entity/enums/RegistrationSource.java
@@ -1,4 +1,4 @@
-package com.pintogether.backend.domain.enums;
+package com.pintogether.backend.entity.enums;
 
 public enum RegistrationSource {
